requirements.txt
appdirs==1.4.3
futures==3.1.1
netaddr==0.7.19
packaging==16.8
pbkdf2==1.3
Pillow==4.1.0
pycket==0.3.0
PyMySQL==0.7.11
pyparsing==2.2.0
qiniu==7.1.2
redis==2.10.5
requests==2.14.2
SQLAlchemy==1.1.9
tornado==4.4.3
torndb==0.3
